+++ src/main/java/cm/ptks/craftflowers/flower/AgingFlower.java
@@ -1,6 +1,6 @@
 package cm.ptks.craftflowers.flower;
 
-import cm.ptks.craftflowers.storage.LanguageFile;
+import cm.ptks.craftflowers.languages.LanguageFile;
 import com.google.gson.JsonObject;
 import org.bukkit.Material;
