+++ src/main/java/cm/ptks/craftflowers/flower/FlowerRegistry.java
@@ -1,6 +1,6 @@
 package cm.ptks.craftflowers.flower;
 
-import cm.ptks.craftflowers.storage.LanguageFile;
+import cm.ptks.craftflowers.languages.LanguageFile;
 import org.bukkit.Material;
 
 import java.util.ArrayList;
